fix: 댓글 좋아요 여부 표시 코드 수정

+++ TimeTableArtist/src/main/java/SamwaMoney/TimeTableArtist/ReplyLike/controller/ReplyLikeController.java
@@ -16,16 +16,25 @@ public class ReplyLikeController {
     @ResponseStatus(value = HttpStatus.CREATED)
     public String createReplyLike(@PathVariable Long replyId, @RequestBody ReplyLikeRequestDto requestDto){
         Long memberId = requestDto.getMemberId();
-        replyLikeService.create(replyId, memberId);
-        return "좋아요를 눌렀습니다.";
+        if (memberId != null){
+            replyLikeService.create(replyId, memberId);
+            return "좋아요를 눌렀습니다.";
+        }
+        else{
+            return "로그인이 필요한 서비스입니다.";
+        }
     }
 
     @DeleteMapping
     @ResponseStatus(value = HttpStatus.OK)
     public String deleteReplyLike(@PathVariable Long replyId, @RequestBody ReplyLikeRequestDto requestDto){
         Long memberId = requestDto.getMemberId();
-        replyLikeService.delete(replyId, memberId);
-        return "좋아요가 취소되었습니다.";
+        if (memberId != null){
+            replyLikeService.delete(replyId, memberId);
+            return "좋아요가 취소되었습니다.";
+        }
+        else{
+            return "로그인이 필요한 서비스입니다.";
+        }
     }
-
 }
